Revert "Code refactor"
