+++ client/package.json
@@ -8,7 +8,6 @@
     "axios": "^0.16.2",
     "history": "^4.7.2",
     "jwt-decode": "^2.2.0",
-    "prop-types": "^15.6.2",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-materialize": "^1.1.2",
