--- client/src/components/UploadForm/index.js
@@ -1,1 +1,0 @@
-export { default } from "./UploadForm";
